refactor: move styles.js into utils dir
