+++ src/components/Feedback.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FeedbackSnackbar, CircularProgress } from '@dhis2/d2-ui-core';
 import { LOADING } from '../utils/feedbackSnackBarTypes';
-import styles from '../styles';
+import styles from '../utils/styles';
 
 const Feedback = ({ open, conf, onClose }) => (
     conf.type === LOADING
