+++ src/pages/data-approval/DataApproval.js
@@ -29,7 +29,7 @@ import { getDocsUrl } from '../../utils/getDocsUrl';
 import { LOADING, SUCCESS } from '../../utils/feedbackSnackBarTypes';
 
 /* styles */
-import styles from '../../styles';
+import styles from '../../utils/styles';
 
 export default class DataApproval extends Page {
     static propTypes = {
