+++ src/pages/organisation-unit-distribution-report/OrganisationUnitDistributionReport.js
@@ -31,7 +31,7 @@ import { getDocsUrl } from '../../utils/getDocsUrl';
 import { LOADING, SUCCESS } from '../../utils/feedbackSnackBarTypes';
 
 /* styles */
-import styles from '../../styles';
+import styles from '../../utils/styles';
 
 export default class OrganisationUnitDistributionReport extends Page {
     static propTypes = {
