+++ src/pages/reporting-rate-summary/ReportingRateSummary.js
@@ -33,7 +33,7 @@ import { getDocsUrl } from '../../utils/getDocsUrl';
 import { LOADING, SUCCESS } from '../../utils/feedbackSnackBarTypes';
 
 /* styles */
-import styles from '../../styles';
+import styles from '../../utils/styles';
 
 const BASED_ON_OPTIONS = [
     {
