+++ src/pages/standard-report/AddReportButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, SvgIcon } from '@dhis2/d2-ui-core';
-import appStyles from '../../styles';
+import appStyles from '../../utils/styles';
 
 const AddReportButton = ({ onClick }) => (
     <div id="add-std-report-btn-container-id">
