+++ src/pages/standard-report/create-report/CreateStdReport.js
@@ -18,7 +18,7 @@ import { i18nKeys } from '../../../utils/i18n/i18nKeys';
 import i18n from '../../../utils/i18n/locales';
 
 /* styles */
-import appStyles from '../../../styles';
+import appStyles from '../../../utils/styles';
 
 import { REPORTS_ENDPOINT, REPORT_MODE, GET_REPORT_AS_ENDPOINT, TYPES } from '../standard.report.conf';
